RPC, cosmetic: move more RPC code to new rpcblockchain.cpp module
